Add files via upload
Switches_fish_fry_liquid_feeder_peristaltic_pump_CNC_shield.ino

Video soon as I upload it

Ardiuno Uno
CNC_shield V3
NEMA 17 stepper. In my case the plan is to drive a Peristaltic pump for feeding fish fry liquid ever 4 hours

It has manual move stepper buttons with LED signals

This code is for a Peristaltic pump for feeding fish fry liquid ever 4 hours.

The code here has the time and the total distance moved reduced for testing purposes.

It was originally for use with a syringe; this in this code below. I ask for a return to the start position. This is most likely not required for use or even recommended for use with a Peristaltic pump.

Both a Peristaltic pump or a Syringe will require a small amount of back movement, (not in this code) to lessen any drops of liquid.

I have issues with both liquid adhesion and cohesion, that after a drop is pushed out this caused more liquid to be pulled out. One maybe possible way of helping prevent this maybe, (still testing at this time), is to put the end of the syringe needle or Peristaltic pump tube end in to the actual fish tank where the baby fry are.

This only switches the Stepper control on just before moving and off after moving, this reduces heat but loses holding force, not required for a Peristaltic pump I plan to use.

Uses:
NoDelay library
AccelStepper

Information
   by Dejan, https://howtomechatronics.com
   Ref: https://howtomechatronics.com/tutorials/arduino/stepper-motors-and-arduino-the-ultimate-guide/#TMC2208_Stepper_Driver
   also uses other online code

 The AccelStepper library for Arduino
 http://www.airspayce.com/mikem/arduino/AccelStepper/index.html

 AccelStepper commands
 http://www.airspayce.com/mikem/arduino/AccelStepper/classAccelStepper.html

Based on NoDelay library. This is the Magic Delay option  https://github.com/M-tech-Creations/NoDelay/blob/baad513b9608959136b8c519803c13e470111fd7/examples/BlinkWithNoDelay/BlinkWithNoDelay.ino

The AccelStepper library for Arduino
http://www.airspayce.com/mikem/arduino/AccelStepper/index.html

http://www.airspayce.com/mikem/arduino/AccelStepper/classAccelStepper.html#a748665c3962e66fbc0e9373eb14c69c1

Debugging in Visual Code https://devblogs.microsoft.com/iotdev/debug-your-arduino-code-with-visual-studio-code/

Timer (DID NOT WORK). So use NoDelay
https://www.pjrc.com/teensy/td_libs_TimerOne.html
https://playground.arduino.cc/Code/Timer1/

interrupts (DID NOT WORK). So use NoDelay
https://arduino.stackexchange.com/questions/30968/how-do-interrupts-work-on-the-arduino-uno-and-similar-boards

Do not use a jumper between EN/GND on the CNC board. In code stepper enable is turned on to make a move then turned off after that move, false is on. This is to avoid holding power causing heating. Stepper may be moved by hand because there is no holding power.

On the CNC shield the limit switches are expected to be normally open and are wired to ground with the internal pullup enabled. The switch will read high when not actuated (open) and low when actuated (closed).

Ref: https://www.electroniclinic.com/arduino-cnc-shield-v3-0-and-a4988-hybrid-stepper-motor-driver-joystick/